requirements.txt
praw
requests
python-dotenv
groq
datetime
schedule